+++ public/data/visualData.json
@@ -7,7 +7,7 @@
   {
     "id": 2,
     "alt": "profileimg",
-    "src": "https://www.teahub.io/photos/full/91-918814_abstract-wallpapers-1920-600.jpg"
+    "src": "https://images.unsplash.com/photo-1627933907906-6075f7c88da5?ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&ixlib=rb-1.2.1&auto=format&fit=crop&w=2735&q=80"
   },
   {
     "id": 3,
