+++ src/pages/Main/ImgCom.js
@@ -3,11 +3,7 @@ import React, { Component } from 'react';
 class ImgCom extends Component {
   render() {
     const { src, alt } = this.props;
-    return (
-      <>
-        <img src={src} alt={alt}></img>
-      </>
-    );
+    return <img src={src} alt={alt} />;
   }
 }
