--- src/pages/Main/Main.scss
@@ -1,35 +1,0 @@
-.imgWrap {
-  border: 1px solid red;
-  width: 100%;
-  height: 100vh;
-  display: flex;
-  align-items: center;
-  overflow: hidden;
-}
-.slide {
-  border: 1px solid blue;
-  min-width: 100%;
-  height: 600px;
-  transition: 0.5s;
-}
-img {
-  height: 600px;
-}
-.leftButton {
-  position: absolute;
-  top: 50%;
-  left: 0px;
-  transform: translateY(-50%);
-  width: 50px;
-  height: 50px;
-  background: none;
-}
-.rightButton {
-  position: absolute;
-  top: 50%;
-  right: 0px;
-  transform: translateY(-50%);
-  width: 50px;
-  height: 50px;
-  background: none;
-}
