+++ src/pages/Main/MainCarousel/MainCarousel.scss
@@ -1,0 +1,64 @@
+.visualWrap {
+  width: 1000px;
+  height: 600px;
+  background-color: blue;
+
+  .visualSlide {
+    width: 1000px;
+    height: 100%;
+
+    .imgWrap {
+      display: flex;
+
+      .slide {
+        width: 100%;
+        height: 600px;
+        border: 1px solid black;
+        background-color: red;
+      }
+    }
+  }
+}
+
+// .visualSlide {
+//   width: 500px;
+//   height: 600px;
+//   background-color: red;
+//   overflow: hidden;
+// }
+
+// .imgWrap {
+//   display: flex;
+//   align-items: center;
+//   width: 100%;
+//   height: 100vh;
+
+//   .slide {
+//     border: 1px solid blue;
+//     width: 500px;
+//     height: 600px;
+//     transition: 0.5s;
+
+//     img {
+//       height: 600px;
+//     }
+//   }
+// }
+.leftButton {
+  position: absolute;
+  top: 50%;
+  left: 0px;
+  transform: translateY(-50%);
+  width: 50px;
+  height: 50px;
+  background: none;
+}
+.rightButton {
+  position: absolute;
+  top: 50%;
+  right: 0px;
+  transform: translateY(-50%);
+  width: 50px;
+  height: 50px;
+  background: none;
+}
